Added contracts for all remaining entities Added services folder structure for all remaining enttities Completed CQRS for AssetProduct  - Commands (Create, Update, Delete)  - Queries(GetById, GetAll)

+++ ITAssetTracker.Application/Services/Assets/Commands/DeleteAsset/DeleteAssetCommandHandler.cs
@@ -1,27 +1,30 @@
 ﻿using AutoMapper;
 using ITAssetTracker.Application.Contracts.Persistence;
+using ITAssetTracker.Application.Exceptions;
 using ITAssetTracker.Domain.Entities;
 using MediatR;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace ITAssetTracker.Application.Services.Assets.Commands.DeleteAsset
+namespace ITAssetTracker.Application.Services.Assets.Commands.DeleteAsset;
+
+public class DeleteAssetCommandHandler : IRequestHandler<DeleteAssetCommand>
 {
-    public class DeleteAssetCommandHandler : IRequestHandler<DeleteAssetCommand>
-    {
-        private readonly IAsyncRepository<Asset> assetRepository;
-        private readonly IMapper mapper;
+    private readonly IAsyncRepository<Asset> assetRepository;
+    private readonly IMapper mapper;
 
-        public DeleteAssetCommandHandler(IAsyncRepository<Asset> assetRepository, IMapper mapper)
+    public DeleteAssetCommandHandler(IAsyncRepository<Asset> assetRepository, IMapper mapper)
+    {
+        this.assetRepository = assetRepository;
+        this.mapper = mapper;
+    }
+    public async Task Handle(DeleteAssetCommand request, CancellationToken cancellationToken)
+    {
+        Asset? assetToDelete = await assetRepository.GetByIdAsync(request.Id);
+        
+        if (assetToDelete is null)
         {
-            this.assetRepository = assetRepository;
-            this.mapper = mapper;
+            throw new NotFoundException("Asset", request.Id);
         }
-        public async Task Handle(DeleteAssetCommand request, CancellationToken cancellationToken)
-        {
-            Asset assetToDelete = await assetRepository.GetByIdAsync(request.Id);
-            await assetRepository.DeleteAsync(assetToDelete);
-        }
+        
+        await assetRepository.DeleteAsync(assetToDelete);
     }
 }
